+++ src/services/ConversationService.js
@@ -8,17 +8,47 @@ import ConversationMember from '../DTOs/ConversationMember';
 import Conversation from '../models/Conversation';
 import APIEndpoints from '../enum/APIEndpoints';
 import CryptoUtils from '../utils/CryptoUtils';
+import MessageService from './MessageService';
 import Injector from '../facades/Injector';
 import Request from '../facades/Request';
 import UserService from './UserService';
 import Service from './Service';
-import conversation from '../models/Conversation';
+
+/**
+ * @typedef ConversationProperties
+ *
+ * @property {AESStaticParametersProperties} encryptionParameters
+ * @property {HMACSigningParametersProperties} signingParameters
+ * @property {ConversationMemberProperties[]} members
+ * @property {?string} name
+ * @property {string} id
+ */
 
 class ConversationService extends Service {
+    /**
+     * @type {ConversationRepository}
+     */
     #conversationRepository;
-    #webSocketClient;
-    #conversation;
 
+    /**
+     * @type {WebSocketClient}
+     */
+    #webSocketClient;
+
+    /**
+     * @type {?Conversation}
+     */
+    #conversation = null;
+
+    /**
+     * Builds a conversation member placeholder instance.
+     *
+     * @param {User} user
+     * @param {string} exportedEncryptionKey
+     * @param {string} exportedSigningKey
+     *
+     * @returns {Promise<ConversationMemberPlaceholder>}
+     */
     async #buildConversationMemberPlaceholder(user, exportedEncryptionKey, exportedSigningKey){
         const RSAPublicKey = await CryptoUtils.importRSAPublicKey(user.getRSAPublicKey());
         const serializedEncryptionKey = JSON.stringify(exportedEncryptionKey);
@@ -32,6 +62,13 @@ class ConversationService extends Service {
         });
     }
 
+    /**
+     * Looks up and assigns user model instances to each conversation member.
+     *
+     * @param {Conversation} conversation
+     *
+     * @returns {Promise<void>}
+     */
     async #assignUsersToConversation(conversation){
         const userIDList = new Set(), userService = new UserService();
         conversation.getMembers().forEach((member) => userIDList.add(member.userID));
@@ -54,6 +91,13 @@ class ConversationService extends Service {
         }
     }
 
+    /**
+     * Stores a conversation locally given its properties.
+     *
+     * @param {ConversationProperties} conversationProperties
+     *
+     * @returns {Promise<Conversation>}
+     */
     async #storeSingleConversation(conversationProperties){
         const { id, encryptionParameters, signingParameters, name } = conversationProperties, members = [], userList = [];
         const hmacSigningParameters = new HMACSigningParameters(signingParameters);
@@ -73,6 +117,11 @@ class ConversationService extends Service {
         return conversation;
     }
 
+    /**
+     * The class constructor.
+     *
+     * @param {?Conversation} [conversation]
+     */
     constructor(conversation = null){
         super();
 
@@ -81,23 +130,53 @@ class ConversationService extends Service {
         this.setConversation(conversation);
     }
 
+    /**
+     * Sets the conversation.
+     *
+     * @param {?Conversation} conversation
+     *
+     * @returns {ConversationService}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation is given.
+     */
     setConversation(conversation){
         if ( conversation !== null && !( conversation instanceof Conversation ) ){
-            throw new IllegalArgumentException('Invalid conversation instance.');
+            throw new IllegalArgumentException('Invalid conversation.');
         }
         this.#conversation = conversation;
         return this;
     }
 
+    /**
+     * Returns the conversation.
+     *
+     * @returns {?Conversation}
+     */
     getConversation(){
         return this.#conversation;
     }
 
-    async storeConversation(properties){
-        this.#conversation = await this.#storeSingleConversation(properties);
-        return this.#conversation;
+    /**
+     * Stores a conversation locally given its properties.
+     *
+     * @param {ConversationProperties} conversationProperties
+     *
+     * @returns {Promise<Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation properties object is given.
+     */
+    async storeConversation(conversationProperties){
+        if ( conversationProperties === null || typeof conversationProperties !== 'object' ){
+            throw new IllegalArgumentException('Invalid conversation properties.');
+        }
+        return this.#conversation = await this.#storeSingleConversation(conversationProperties);
     }
 
+    /**
+     * Fetches and stores conversations locally.
+     *
+     * @returns {Promise<Conversation[]>}
+     */
     async fetchConversations(){
         const response = await Request.get(APIEndpoints.CONVERSATION_LIST, null, true);
         const conversationList = await Promise.all(response.conversationList.map((conversationProperties) => {
@@ -107,6 +186,11 @@ class ConversationService extends Service {
         return conversationList;
     }
 
+    /**
+     * Returns all the locally stored conversations.
+     *
+     * @returns {Promise<Conversation[]>}
+     */
     async getConversations(){
         const conversationList = await this.#conversationRepository.getAll();
         await Promise.all(conversationList.map((conversation) => {
@@ -115,7 +199,19 @@ class ConversationService extends Service {
         return conversationList;
     }
 
+    /**
+     * Returns a conversation given its ID.
+     *
+     * @param {string} conversationID
+     *
+     * @returns {Promise<?Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation ID is given.
+     */
     async getConversationByID(conversationID){
+        if ( conversationID === '' || typeof conversationID !== 'string' ){
+            throw new IllegalArgumentException('Invalid conversation ID.');
+        }
         this.#conversation = await this.#conversationRepository.getByID(conversationID);
         if ( this.#conversation !== null ){
             await this.#assignUsersToConversation(this.#conversation);
@@ -123,7 +219,19 @@ class ConversationService extends Service {
         return this.#conversation;
     }
 
+    /**
+     * Fetches and then stores a conversation given its ID.
+     *
+     * @param {string} conversationID
+     *
+     * @returns {Promise<?Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation ID is given.
+     */
     async fetchConversationByID(conversationID){
+        if ( conversationID === '' || typeof conversationID !== 'string' ){
+            throw new IllegalArgumentException('Invalid conversation ID.');
+        }
         const url = APIEndpoints.CONVERSATION_GET.replace(':conversationID', conversationID);
         const response = await Request.get(url, null, true);
         this.#conversation = await this.#storeSingleConversation(response.conversation);
@@ -131,7 +239,19 @@ class ConversationService extends Service {
         return this.#conversation;
     }
 
+    /**
+     * Returns a conversation given its ID, if not stored locally it will be fetched from the server.
+     *
+     * @param {string} conversationID
+     *
+     * @returns {Promise<?Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation ID is given.
+     */
     async assertConversation(conversationID){
+        if ( conversationID === '' || typeof conversationID !== 'string' ){
+            throw new IllegalArgumentException('Invalid conversation ID.');
+        }
         this.#conversation = await this.getConversationByID(conversationID);
         if ( this.#conversation === null ){
             await this.fetchConversationByID(conversationID);
@@ -139,7 +259,24 @@ class ConversationService extends Service {
         return this.#conversation;
     }
 
+    /**
+     * Returns the direct message conversation where the given members are involved in.
+     *
+     * @param {string} senderUserID
+     * @param {string} recipientUserID
+     *
+     * @returns {Promise<?Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid recipient user ID is given.
+     * @throws {IllegalArgumentException} If an invalid sender user ID is given.
+     */
     async getDMConversationByMembers(senderUserID, recipientUserID){
+        if ( recipientUserID === '' || typeof recipientUserID !== 'string' ){
+            throw new IllegalArgumentException('Invalid recipient user ID.');
+        }
+        if ( senderUserID === '' || typeof senderUserID !== 'string' ){
+            throw new IllegalArgumentException('Invalid sender user ID.');
+        }
         let conversationList = await this.#conversationRepository.getAll(), DMConversation = null, i = 0;
         while ( DMConversation === null && i < conversationList.length ){
             const members = conversationList[i].getMembers();
@@ -158,20 +295,29 @@ class ConversationService extends Service {
     /**
      * Creates a new conversation.
      *
-     * @param {User} userList
-     * @param {string} conversationName
+     * @param {User[]} memberList
+     * @param {?string} [conversationName]
      *
      * @returns {Promise<Conversation>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation name is given.
+     * @throws {IllegalArgumentException} If an invalid member list is given.
      */
-    async createConversation(userList, conversationName = null){
+    async createConversation(memberList, conversationName = null){
+        if ( conversationName !== null && ( conversationName === '' || typeof conversationName !== 'string' ) ){
+            throw new IllegalArgumentException('Invalid conversation name.');
+        }
+        if ( !Array.isArray(memberList) ){
+            throw new IllegalArgumentException('Invalid member list.');
+        }
         const aesStaticParameters = new AESStaticParameters({ keyLength: 256, mode: 'GCM' });
         const hmacSigningParameters = new HMACSigningParameters({ hashName: 'SHA-512' });
         const signingKey = await CryptoUtils.generateHMACKey(hmacSigningParameters);
         const encryptionKey = await CryptoUtils.generateAESKey(aesStaticParameters);
         const exportedEncryptionKey = await CryptoUtils.exportKey(encryptionKey);
         const exportedSigningKey = await CryptoUtils.exportKey(signingKey);
-        const conversationMemberPlaceholderList = await Promise.all(userList.map((user) => {
-            return this.#buildConversationMemberPlaceholder(user, exportedEncryptionKey, exportedSigningKey);
+        const conversationMemberPlaceholderList = await Promise.all(memberList.map((member) => {
+            return this.#buildConversationMemberPlaceholder(member, exportedEncryptionKey, exportedSigningKey);
         }));
         const response = await Request.post(APIEndpoints.CONVERSATION_CREATE, {
             conversationMemberPlaceholderList: conversationMemberPlaceholderList,
@@ -180,10 +326,14 @@ class ConversationService extends Service {
             encryptionMode: aesStaticParameters.getMode(),
             name: conversationName
         }, true);
-        this.#conversation = await this.#storeSingleConversation(response.conversation);
-        return this.#conversation;
+        return this.#conversation = await this.#storeSingleConversation(response.conversation);
     }
 
+    /**
+     * Notifies the server that the user is typing in the defined conversation.
+     *
+     * @returns {Promise<void>}
+     */
     async notifyTyping(){
         await this.#webSocketClient.send({
             payload: { conversationID: this.#conversation.getID() },
@@ -191,16 +341,45 @@ class ConversationService extends Service {
         });
     }
 
+    /**
+     * Deletes a conversation given its ID.
+     *
+     * @param {string} conversationID
+     *
+     * @returns {Promise<void>}
+     *
+     * @throws {IllegalArgumentException} If an invalid conversation ID is given.
+     */
     async deleteConversationByID(conversationID){
+        if ( conversationID === '' || typeof conversationID !== 'string' ){
+            throw new IllegalArgumentException('Invalid conversation ID.');
+        }
         await this.#conversationRepository.deleteByID(conversationID);
         this._eventBroker.emit('conversationDelete', conversationID);
     }
 
+    /**
+     * Deletes the defined conversation.
+     *
+     * @param {boolean} deleteForEveryone
+     *
+     * @returns {Promise<void>}
+     */
     async delete(deleteForEveryone){
         const url = APIEndpoints.CONVERSATION_DELETE.replace(':conversationID', this.#conversation.getID());
         await Request.delete(url, { deleteForEveryone: ( deleteForEveryone === true ? '1' : '0' ) }, true);
         await this.#conversationRepository.delete(this.#conversation);
         this._eventBroker.emit('conversationDelete', this.#conversation.getID());
+    }
+
+    /**
+     * Marks every message contained in the conversation defined as read.
+     *
+     * @returns {Promise<void>}
+     */
+    async markAsRead(){
+        await Request.patch(APIEndpoints.CONVERSATION_MARK_AS_READ.replace(':conversationID', this.#conversation.getID()));
+        await new MessageService(this.#conversation).markMessagesAsRead();
     }
 }
